More fixes for HDF5 1.8 and CentOS with older Boost 1.41.
Some of the attribute functions are unimplemented, and a message will occur during compilation is these definitions need to be filled in for the code to work. Happens in 2-3 places.

Bug in Boost Karma 1.41 prevents app build. For simplicity, I think that I will write a custom shappefile parser / will import from an older version of rtmath. This will be slower, but the example doesn't really need the fastest parser, anyways.

Need to also add in unit tests of the different functions, and need to add substantial documentation. Need to reimplement any missing read functions, too.