Pausing for guidance on exvs. The writing code gets called.
Will branch the code so that I can work on extending the shapes implementation to accept plugins. DDSCAT shapefile import will be extended. I really want to be able to 'pipeline' the shapes, and want to be able to export the shapes to different 3-D formats. DLL code cleanup, too.